2018 Day 9 part 1
I mis read the instructions and spent wasted time. The last marble is worth is the marble number not the last scored value
Part 1 is done and I attempted part 2 but it has 7.1e6 marbles so I got a bad highscore back and it took 2 + hours. I need to look at switching to the std::deque and see what happens when the speed improves.