Part 2 solved
I was on the right track but let the documentation fool me. I wanted to --it and ++it but thought I could not with the standard iterator. When I used the code from : https://github.com/Chrinkus/advent-of-code-2018/blob/master/src/day09.cpp for incrementing and decrementing it worked great.